3163 refactor some base mdm stuff
