+++ hapi-fhir-base/src/main/java/org/hl7/fhir/instance/model/api/IAnyResource.java
@@ -31,7 +31,7 @@ public interface IAnyResource extends IBaseResource {
 	/**
 	 * Search parameter constant for <b>_id</b>
 	 */
-	@SearchParamDefinition(name="_id", path="", description="The ID of the resource", type="token"  )
+	@SearchParamDefinition(name="_id", path="", description="The ID of the resource", type="token")
 	String SP_RES_ID = "_id";
 
 	/**
