+++ hapi-fhir-jpaserver-mdm/src/test/java/ca/uhn/fhir/jpa/mdm/provider/BaseProviderR4Test.java
@@ -5,6 +5,7 @@ import ca.uhn.fhir.mdm.api.IMdmClearJobSubmitter;
 import ca.uhn.fhir.mdm.api.IMdmControllerSvc;
 import ca.uhn.fhir.mdm.api.IMdmMatchFinderSvc;
 import ca.uhn.fhir.mdm.api.IMdmSubmitSvc;
+import ca.uhn.fhir.mdm.provider.MdmControllerHelper;
 import ca.uhn.fhir.mdm.provider.MdmProviderDstu3Plus;
 import ca.uhn.fhir.mdm.rules.config.MdmSettings;
 import ca.uhn.fhir.mdm.util.MessageHelper;
@@ -29,8 +30,6 @@ import java.util.List;
 public abstract class BaseProviderR4Test extends BaseMdmR4Test {
 	MdmProviderDstu3Plus myMdmProvider;
 	@Autowired
-	private IMdmMatchFinderSvc myMdmMatchFinderSvc;
-	@Autowired
 	private IMdmControllerSvc myMdmControllerSvc;
 	@Autowired
 	private IMdmClearJobSubmitter myMdmClearJobSubmitter;
@@ -38,6 +37,8 @@ public abstract class BaseProviderR4Test extends BaseMdmR4Test {
 	private IMdmSubmitSvc myMdmSubmitSvc;
 	@Autowired
 	private MdmSettings myMdmSettings;
+	@Autowired
+	private MdmControllerHelper myMdmHelper;
 	@Autowired
 	BatchJobHelper myBatchJobHelper;
 	@Autowired
@@ -55,7 +56,7 @@ public abstract class BaseProviderR4Test extends BaseMdmR4Test {
 
 	@BeforeEach
 	public void before() {
-		myMdmProvider = new MdmProviderDstu3Plus(myFhirContext, myMdmControllerSvc, myMdmMatchFinderSvc, myMdmSubmitSvc, myMdmSettings);
+		myMdmProvider = new MdmProviderDstu3Plus(myFhirContext, myMdmControllerSvc, myMdmHelper, myMdmSubmitSvc, myMdmSettings);
 		defaultScript = myMdmSettings.getScriptText();
 	}
