+++ hapi-fhir-server-mdm/src/main/java/ca/uhn/fhir/mdm/provider/MdmControllerHelper.java
@@ -21,19 +21,35 @@ package ca.uhn.fhir.mdm.provider;
  */
 
 import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.mdm.api.IMdmMatchFinderSvc;
 import ca.uhn.fhir.mdm.api.IMdmSettings;
+import ca.uhn.fhir.mdm.api.MatchedTarget;
+import ca.uhn.fhir.mdm.api.MdmConstants;
 import ca.uhn.fhir.mdm.util.MdmResourceUtil;
 import ca.uhn.fhir.mdm.util.MessageHelper;
 import ca.uhn.fhir.model.primitive.IdDt;
 import ca.uhn.fhir.rest.server.exceptions.InvalidRequestException;
 import ca.uhn.fhir.rest.server.exceptions.ResourceVersionConflictException;
 import ca.uhn.fhir.rest.server.provider.ProviderConstants;
+import ca.uhn.fhir.util.BundleBuilder;
 import ca.uhn.fhir.validation.IResourceLoader;
 import org.hl7.fhir.instance.model.api.IAnyResource;
+import org.hl7.fhir.instance.model.api.IBase;
+import org.hl7.fhir.instance.model.api.IBaseBackboneElement;
+import org.hl7.fhir.instance.model.api.IBaseBundle;
+import org.hl7.fhir.instance.model.api.IBaseDatatype;
+import org.hl7.fhir.instance.model.api.IBaseExtension;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.instance.model.api.IIdType;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import javax.annotation.Nonnull;
+import java.math.BigDecimal;
+import java.util.Comparator;
+import java.util.Date;
+import java.util.List;
+import java.util.UUID;
 
 @Service
 public class MdmControllerHelper {
@@ -42,12 +58,18 @@ public class MdmControllerHelper {
 	private final IResourceLoader myResourceLoader;
 	private final IMdmSettings myMdmSettings;
 	private final MessageHelper myMessageHelper;
+	private final IMdmMatchFinderSvc myMdmMatchFinderSvc;
 
 	@Autowired
-	public MdmControllerHelper(FhirContext theFhirContext, IResourceLoader theResourceLoader, IMdmSettings theMdmSettings, MessageHelper theMessageHelper) {
+	public MdmControllerHelper(FhirContext theFhirContext,
+										IResourceLoader theResourceLoader,
+										IMdmMatchFinderSvc theMdmMatchFinderSvc,
+										IMdmSettings theMdmSettings,
+										MessageHelper theMessageHelper) {
 		myFhirContext = theFhirContext;
 		myResourceLoader = theResourceLoader;
 		myMdmSettings = theMdmSettings;
+		myMdmMatchFinderSvc = theMdmMatchFinderSvc;
 		myMessageHelper = theMessageHelper;
 	}
 
@@ -103,5 +125,59 @@ public class MdmControllerHelper {
 		if (!MdmResourceUtil.isMdmManaged(theResource)) {
 			throw new InvalidRequestException(myMessageHelper.getMessageForUnmanagedResource());
 		}
+	}
+
+	/**
+	 * Helper method which will return a bundle of all Matches and Possible Matches.
+	 */
+	public IBaseBundle getMatchesAndPossibleMatchesForResource(IAnyResource theResource, String theResourceType) {
+		List<MatchedTarget> matches = myMdmMatchFinderSvc.getMatchedTargets(theResourceType, theResource);
+		matches.sort(Comparator.comparing((MatchedTarget m) -> m.getMatchResult().getNormalizedScore()).reversed());
+
+		BundleBuilder builder = new BundleBuilder(myFhirContext);
+		builder.setBundleField("type", "searchset");
+		builder.setBundleField("id", UUID.randomUUID().toString());
+		builder.setMetaField("lastUpdated", builder.newPrimitive("instant", new Date()));
+
+		IBaseBundle retVal = builder.getBundle();
+		for (MatchedTarget next : matches) {
+			boolean shouldKeepThisEntry = next.isMatch() || next.isPossibleMatch();
+			if (!shouldKeepThisEntry) {
+				continue;
+			}
+
+			IBase entry = builder.addEntry();
+			builder.addToEntry(entry, "resource", next.getTarget());
+
+			IBaseBackboneElement search = builder.addSearch(entry);
+			toBundleEntrySearchComponent(builder, search, next);
+		}
+		return retVal;
+	}
+
+	public IBaseBackboneElement toBundleEntrySearchComponent(BundleBuilder theBuilder, IBaseBackboneElement theSearch, MatchedTarget theMatchedTarget) {
+		theBuilder.setSearchField(theSearch, "mode", "match");
+		double score = theMatchedTarget.getMatchResult().getNormalizedScore();
+		theBuilder.setSearchField(theSearch, "score",
+			theBuilder.newPrimitive("decimal", BigDecimal.valueOf(score)));
+
+		String matchGrade = getMatchGrade(theMatchedTarget);
+		IBaseDatatype codeType = (IBaseDatatype) myFhirContext.getElementDefinition("code").newInstance(matchGrade);
+		IBaseExtension searchExtension = theSearch.addExtension();
+		searchExtension.setUrl(MdmConstants.FIHR_STRUCTURE_DEF_MATCH_GRADE_URL_NAMESPACE);
+		searchExtension.setValue(codeType);
+
+		return theSearch;
+	}
+
+	@Nonnull
+	protected String getMatchGrade(MatchedTarget theTheMatchedTarget) {
+		String retVal = "probable";
+		if (theTheMatchedTarget.isMatch()) {
+			retVal = "certain";
+		} else if (theTheMatchedTarget.isPossibleMatch()) {
+			retVal = "possible";
+		}
+		return retVal;
 	}
 }
